Added new fields to identify host and insertion material ids
I added 2 new fields so that the StructureGroupDoc returns the host and insertion material ids. The host material id corresponds to the lowest concentration of working ion structure in the group (and lowest energy if multiple structures share the lowest concentration). The insertion material ids correspond to the remaining higher concentration structures. The StructureGroupDoc also allowed the user to specify a list of "ignored species" but based on how it is used in the StructureGroupBuilder, it makes more sense to specify its type as a string so corresponding changes were made to make this consistent.