[WIP] Updated StructureGroupDoc (#491)
* Added new fields to identify host and insertion material ids

I added 2 new fields so that the StructureGroupDoc returns the host and insertion material ids. The host material id corresponds to the lowest concentration of working ion structure in the group (and lowest energy if multiple structures share the lowest concentration). The insertion material ids correspond to the remaining higher concentration structures. The StructureGroupDoc also allowed the user to specify a list of "ignored species" but based on how it is used in the StructureGroupBuilder, it makes more sense to specify its type as a string so corresponding changes were made to make this consistent.

* corrected input error for StructureMatcher

"ignored_species" should be a list containing the single "ignored_specie" when fed into the StructureMatcher

* corrected calling for field from StructureMatcher

"ignored_specie" is not the correct field to call

* Update how working ion is called from the StructureMatcher

* get energy data from material doc to be able to compare energies among host structures

* updated the document model fields and fixed how the host structure is identified

* reformatted code

* reformatted code

* reformatted code

* added MP compatibility corrections

* updated group_id and host_ids fields

* update structure_group test

* updated structure_group test assertion